+++ fast64_internal/oot/importer/scene_collision.py
@@ -2,16 +2,17 @@ import math
 import re
 import bpy
 import mathutils
+
 from random import random
 from collections import OrderedDict
-from ....utility import PluginError, parentObject, hexOrDecInt, yUpToZUp
-from ...collision.properties import OOTMaterialCollisionProperty
-from ...oot_f3d_writer import getColliderMat
-from ...oot_utility import setCustomProperty, ootParseRotation
+from ...utility import PluginError, parentObject, hexOrDecInt, yUpToZUp
+from ..collision.properties import OOTMaterialCollisionProperty
+from ..oot_f3d_writer import getColliderMat
+from ..oot_utility import setCustomProperty, ootParseRotation
 from .utility import getDataMatch, getBits, checkBit, createCurveFromPoints, stripName
 from .classes import SharedSceneData
 
-from ...collision.constants import (
+from ..collision.constants import (
     ootEnumFloorSetting,
     ootEnumWallSetting,
     ootEnumFloorProperty,
