+++ fast64_internal/oot/importer/scene_header.py
@@ -4,11 +4,11 @@ import re
 import bpy
 import mathutils
 
-from ....utility import PluginError, readFile, parentObject, hexOrDecInt, gammaInverse
-from ....f3d.f3d_parser import parseMatrices
-from ...oot_model_classes import OOTF3DContext
-from ...scene.properties import OOTSceneHeaderProperty, OOTLightProperty
-from ...oot_utility import getEvalParams, setCustomProperty
+from ...utility import PluginError, readFile, parentObject, hexOrDecInt, gammaInverse
+from ...f3d.f3d_parser import parseMatrices
+from ..oot_model_classes import OOTF3DContext
+from ..scene.properties import OOTSceneHeaderProperty, OOTLightProperty
+from ..oot_utility import getEvalParams, setCustomProperty
 from .constants import headerNames
 from .utility import getDataMatch, stripName
 from .classes import SharedSceneData
@@ -17,7 +17,7 @@ from .actor import parseTransActorList, parseSpawnList, parseEntranceList
 from .scene_collision import parseCollisionHeader
 from .scene_pathways import parsePathList
 
-from ...oot_constants import (
+from ..oot_constants import (
     ootEnumAudioSessionPreset,
     ootEnumNightSeq,
     ootEnumMusicSeq,
