+++ fast64_internal/oot/importer/utility.py
@@ -2,9 +2,9 @@ import re
 import bpy
 import mathutils
 
-from ....utility import PluginError, hexOrDecInt, removeComments, yUpToZUp
-from ...actor.properties import OOTActorProperty, OOTActorHeaderProperty
-from ...oot_utility import ootParseRotation
+from ...utility import PluginError, hexOrDecInt, removeComments, yUpToZUp
+from ..actor.properties import OOTActorProperty, OOTActorHeaderProperty
+from ..oot_utility import ootParseRotation
 from .constants import headerNames, actorsWithRotAsParam
